+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Develop
 
+## v1.1.0
+
+- Print error message only on one trial
+- Add print option to list all expired watchdogs
+
 ## v1.0.0
 
 - Added option to remove wdg from the list